Integrated libphoto2 into photo_node for ROS. Basic testing complete, libphoto2 giving similar results to libphoto.
git-svn-id: http://svn.code.sf.net/p/bosch-ros-pkg/code/trunk@1493 2f5dae86-743c-430d-921f-7dbcfe3428eb

+++ include/photo/photo_camera.hpp
@@ -71,7 +71,7 @@ public:
 
 
   //* Open the 'n'th photo_camera in the list
-  bool photo_camera_open( photo_camera_list* list, size_t n );
+  bool photo_camera_open( photo_camera_list* list, int n );
 
   //* Open a connection the photo_camera of 'model' on port 'port'.
   bool photo_camera_open( photo_camera_list* list, const std::string model_name, std::string port_name );
